requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click
colorama==0.4.6
contourpy
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.122.0
fonttools==4.60.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inference-sdk==0.62.2
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver
langchain-core
langchain-openai
langsmith
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib
multidict==6.7.0
mypy_extensions==1.1.0
numpy
openai==2.8.1
opencv-python==4.10.0.84
orjson==3.11.4
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
propcache==0.4.1
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
scipy
six==1.17.0
sniffio==1.3.1
starlette
supervision==0.27.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
pandas
openpyxl
torch
torchvision
easyocr
llama-parse
nest-asyncio
apscheduler
pymupdf

# Auth / database
SQLAlchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
